Reading last_used as ClientDetails column
