--- iam-login-service/src/main/java/it/infn/mw/iam/api/client/last_used/service/ClientLastUsedService.java
@@ -1,31 +1,0 @@
-/**
- * Copyright (c) Istituto Nazionale di Fisica Nucleare (INFN). 2016-2023
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package it.infn.mw.iam.api.client.last_used.service;
-
-import javax.validation.constraints.NotBlank;
-
-import org.springframework.data.domain.Pageable;
-
-import it.infn.mw.iam.api.common.ListResponseDTO;
-import it.infn.mw.iam.persistence.model.IamClientLastUsed;
-
-public interface ClientLastUsedService {
-
-    ListResponseDTO<IamClientLastUsed> getAllClients(Pageable pageable);
-
-    //void updateLastUsed(@NotBlank String clientId);
-
-}
