--- iam-login-service/src/main/java/it/infn/mw/iam/persistence/model/IamClientLastUsed.java
@@ -1,102 +1,0 @@
-/**
- * Copyright (c) Istituto Nazionale di Fisica Nucleare (INFN). 2016-2023
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package it.infn.mw.iam.persistence.model;
-
-import java.util.Date;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-/**
- * 
- * This entity is the login-service component as it needs to "see" the ClientDetailsEntity mitreid
- * Entity, which is not accessible in the iam-persistence scope.
- * 
- *
- */
-@Entity
-@Table(name = "iam_client_last_used")
-public class IamClientLastUsed {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false, updatable = false)
-    private Long id;
-
-    @Column(name = "client_id", nullable = false, updatable = false)
-    private Long clientId;
-
-    @Temporal(TemporalType.TIMESTAMP)
-    @Column(name = "last_used", nullable = false, updatable = true)
-    private Date lastUsed;
-
-    public IamClientLastUsed() {
-        // emptyOnPurpose
-    }
-
-    public Long getClientId() {
-        return clientId;
-    }
-
-    public void setClient(Long clientId) {
-        this.clientId = clientId;
-    }
-
-    public Date getLastUsed() {
-        return lastUsed;
-    }
-
-    public void setLastUsed(Date lastUsed) {
-        this.lastUsed = lastUsed;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((clientId == null) ? 0 : clientId.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        IamClientLastUsed other = (IamClientLastUsed) obj;
-        if (clientId == null) {
-            if (other.clientId != null) {
-                return false;
-            }
-        } else if (!clientId.equals(other.clientId)) {
-            return false;
-        }
-        return true;
-    }
-
-}
