+++ iam-persistence/src/main/resources/db/migration/h2/V97__client_last_used.sql
@@ -1,7 +1,2 @@
 
-CREATE TABLE IF NOT EXISTS iam_client_last_used (
-  id BIGINT GENERATED BY DEFAULT AS IDENTITY(START WITH 1) PRIMARY KEY,
-  client_id BIGINT NOT NULL,
-  last_used TIMESTAMP NOT NULL);
-
-ALTER TABLE iam_client_last_used ADD CONSTRAINT fk_iam_client_last_used_client_id FOREIGN KEY (client_id) REFERENCES client_details(id);
+ALTER TABLE client_details ADD last_used DATE DEFAULT NULL;
