+++ iam-persistence/src/main/resources/db/migration/test/V100000___test_data.sql
@@ -1530,8 +1530,7 @@ insert into iam_totp_mfa(active, secret, creation_time, last_update_time, accoun
 (true, 'secret', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(), 1000);
 
 -- Client last used dates
-insert into iam_client_last_used(id, client_id, last_used) VALUES
-(1, 1, '1994-03-21'),
-(2, 2, '1994-03-20'),
-(3, 3, '1994-03-19'),
-(4, 4, '1994-03-23');
+UPDATE client_details SET last_used = '1994-03-21' WHERE id = 1;
+UPDATE client_details SET last_used = '1994-03-20' WHERE id = 2;
+UPDATE client_details SET last_used = '1994-03-19' WHERE id = 3;
+UPDATE client_details SET last_used = '1994-03-23' WHERE id = 4;
